2-langchain/requirements.txt
langchain==1.2.17
langchain-openai==1.2.1
langchain-google-genai==4.2.2
python-dotenv==0.19.2